refactor: Extract detail info data request to fnc
