+++ pages/index.tsx
@@ -9,10 +9,11 @@ import MovieCard from "../component/MovieCard";
 import MovieDetail, { Props as DetailProps } from "../component/MovieDetail";
 import Header from "../component/Header";
 import Spinner from "../component/Spinner";
+import { movieService } from "../service/movieService";
 
 const Home: NextPage = () => {
   const defaultDetailState: DetailProps = {
-    info: { content: "", wikiLink: "", imdbLink: "" },
+    info: { content: "", wikiLink: "", imdbLink: "", id: 0 },
     title: "",
     imgSrc: "",
     score: 0,
@@ -72,36 +73,19 @@ const Home: NextPage = () => {
   async function showMovieDetails(
     movieTitle: string,
     score: number,
-    imageSrc: string
+    imageSrc: string,
+    tmdbMovieId: number
   ) {
     showSpinner();
     preventBodyScroll();
-    const result = { content: "", wikiLink: "", imdbLink: "" };
-    const pageDetails = await dataHandler.getWikiPage(movieTitle);
-    const movieId: string = Object.keys(pageDetails.query.pages)[0];
-    if (movieId !== "-1") {
-      if (pageDetails.query.pages[movieId].hasOwnProperty("extract")) {
-        result.content = pageDetails.query.pages[movieId].extract;
-      }
-      result.wikiLink = `https://en.wikipedia.org/wiki/${movieTitle}`;
-      const pageLinks = await dataHandler.getWikiPageLinks(movieTitle);
-      const allLinks = pageLinks.query.pages[movieId].extlinks;
-      if (pageLinks.query.pages[movieId].hasOwnProperty("extlinks")) {
-        const myLinkObj = allLinks.find((current: Object) =>
-          Object.values(current)[0].includes("imdb.com/title")
-        );
-        if (myLinkObj) {
-          result.imdbLink = Object.values(myLinkObj)[0] as string;
-        }
-      }
-    }
-    const newContent = {
-      ...detailsContent,
-      info: { ...result },
-      title: movieTitle,
-      imgSrc: imageSrc,
-      score: score,
-    };
+    const newContent = await movieService.getMovieDetailsData(
+      movieTitle,
+      score,
+      imageSrc,
+      tmdbMovieId,
+      detailsContent
+    );
+    console.log(newContent);
     setDetailsContent({ ...defaultDetailState });
     setDetailsContent(newContent);
     setDetailVisibility("");
@@ -111,7 +95,7 @@ const Home: NextPage = () => {
   return (
     <div className={`${bodyScroll}`}>
       <Head>
-        <title>Create Next App</title>
+        <title>MovieFinder</title>
         <meta name="description" content="Generated by create next app" />
         <link rel="icon" href="/favicon.ico" />
       </Head>
@@ -152,6 +136,7 @@ const Home: NextPage = () => {
                 overview={movie.overview}
                 score={movie.score}
                 key={+movie.id}
+                movieId={+movie.id}
                 fn={showMovieDetails}
               />
             ))}
